backend/requirements.txt
django==5.2.1
django-cors-headers==4.7.0
google-generativeai==0.8.3
python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
